# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Orientamento SaaS API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
      required:
        - error
...
    CalendarEvent:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean }
        category: { type: string, enum: [study, training, interview, deadline, task, follow-up] }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [todo, in-progress, done, postponed] }
        color: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        linkedSectorId: { type: ["integer", "null"] }
        linkedGoal: { type: ["string", "null"] }
        linkedContentIds: { type: array, items: { type: integer } }
        isRecurring: { type: boolean }
        recurrenceRule: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        reminders: { type: array, items: { $ref: "#/components/schemas/EventReminder" } }
      required: [id, userId, title, startAt, endAt, allDay, category, priority, status, createdAt, updatedAt]
    EventReminder:
      type: object
      properties:
        id: { type: integer }
        eventId: { type: integer }
        minutesBefore: { type: integer }
        enabled: { type: boolean }
        sentAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, eventId, minutesBefore, enabled, createdAt]
...
    CreateCalendarEventBody:
      type: object
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: ["string", "null"] }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean, default: false }
        category: { type: string, enum: [study, training, interview, deadline, task, follow-up], default: task }
        priority: { type: string, enum: [low, medium, high], default: medium }
        status: { type: string, enum: [todo, in-progress, done, postponed], default: todo }
        color: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        linkedSectorId: { type: ["integer", "null"] }
        linkedGoal: { type: ["string", "null"] }
        linkedContentIds: { type: array, items: { type: integer } }
        isRecurring: { type: boolean, default: false }
        recurrenceRule: { type: ["string", "null"] }
        reminders:
          type: array
          maxItems: 4
          items:
            type: object
            properties:
              minutesBefore: { type: integer }
              enabled: { type: boolean }
            required: [minutesBefore]
      required: [title, startAt, endAt]
    UpdateCalendarEventBody:
      type: object
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean }
        category: { type: string, enum: [study, training, interview, deadline, task, follow-up] }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [todo, in-progress, done, postponed] }
        color: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        linkedSectorId: { type: ["integer", "null"] }
        linkedGoal: { type: ["string", "null"] }
        linkedContentIds: { type: array, items: { type: integer } }
        reminders:
          type: array
          maxItems: 4
          items:
            type: object
            properties:
              minutesBefore: { type: integer }
              enabled: { type: boolean }
            required: [minutesBefore]
    CalendarQuota:
      type: object
      properties:
        isPremium: { type: boolean }
        eventCount: { type: integer }
        eventLimit: { type: ["integer", "null"] }
      required: [isPremium, eventCount, eventLimit]
...
paths:
...
  /calendar/quota:
    get:
      operationId: getCalendarQuota
      tags: [calendar]
      summary: Get the authenticated user's plan quota and event count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Quota information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarQuota"
  /calendar/events:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List events for the authenticated user within a date range
      security:
        - bearerAuth: []
      parameters:
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: List of events with reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/CalendarEvent" } }
                required: [events]
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Create a new calendar event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCalendarEventBody" }
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: "#/components/schemas/CalendarEvent" }
                required: [event]
        "403":
          description: Free tier limit reached or premium reminder required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /calendar/events/{id}:
    get:
      operationId: getCalendarEvent
      tags: [calendar]
      summary: Get a single event (must own it)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Event with reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: "#/components/schemas/CalendarEvent" }
                required: [event]
        "404": { description: Not found }
    put:
      operationId: updateCalendarEvent
      tags: [calendar]
      summary: Update an event (must own it)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCalendarEventBody" }
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: "#/components/schemas/CalendarEvent" }
                required: [event]
        "403":
          description: Premium reminder required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { description: Not found }
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Delete an event (must own it)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                required: [success]
  /calendar/events/{id}/reminders:
    post:
      operationId: addEventReminder
      tags: [calendar]
      summary: Add a reminder to an event
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minutesBefore: { type: integer }
                enabled: { type: boolean }
              required: [minutesBefore]
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                type: object
                properties:
                  reminder: { $ref: "#/components/schemas/EventReminder" }
                required: [reminder]
        "403":
          description: Premium reminder required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /calendar/reminders/{id}:
    put:
      operationId: updateEventReminder
      tags: [calendar]
      summary: Update a reminder
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minutesBefore: { type: integer }
                enabled: { type: boolean }
      responses:
        "200":
          description: Updated reminder
        "403":
          description: Premium reminder required
    delete:
      operationId: deleteEventReminder
      tags: [calendar]
      summary: Delete a reminder
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
  /calendar/upcoming:
    get:
      operationId: getUpcomingEvents
      tags: [calendar]
      summary: Get upcoming events for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 5, maximum: 20 } }
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/CalendarEvent" } }
                required: [events]
...